+++ edison-status/src/main/java/de/otto/edison/status/configuration/ApplicationStatusAggregatorConfiguration.java
@@ -13,17 +13,16 @@ import java.util.List;
 
 import static java.util.Collections.emptyList;
 
-
 /**
  * Configuration of the default ApplicationStatusAggregator that is used to get and cache the status of this
  * application using all StatusDetailIndicators configured in the Spring application context.
- *
  */
 @Configuration
-public class StatusAggregatorConfiguration {
+public class ApplicationStatusAggregatorConfiguration {
 
     @Autowired(required = false)
     private List<StatusDetailIndicator> statusDetailIndicators = emptyList();
+
     @Autowired(required = false)
     private List<ServiceSpec> serviceSpecs = emptyList();
 
@@ -32,9 +31,9 @@ public class StatusAggregatorConfiguration {
      * {@link de.otto.edison.status.scheduler.Scheduler} every now and then.
      *
      * @param applicationInfo Information about the application
-     * @param versionInfo Information about the application's version
-     * @param systemInfo Information about the system
-     * @param teamInfo Information about the team responsible for this application
+     * @param versionInfo     Information about the application's version
+     * @param systemInfo      Information about the system
+     * @param teamInfo        Information about the team responsible for this application
      * @return ApplicationStatusAggregator
      */
     @Bean