+++ edison-status/src/main/java/de/otto/edison/status/configuration/LoadIndicatorConfiguration.java
@@ -1,17 +1,17 @@
 package de.otto.edison.status.configuration;
 
 import de.otto.edison.status.indicator.load.EverythingFineStrategy;
-import de.otto.edison.status.indicator.load.OverloadDetector;
+import de.otto.edison.status.indicator.load.LoadDetector;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class OverloadIndicatorConfiguration {
+public class LoadIndicatorConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
-    public OverloadDetector defaultStrategy() {
+    public LoadDetector defaultStrategy() {
         return new EverythingFineStrategy();
     }