+++ HISTORY.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## UPCOMING: Release 0.52.0
-* Added status indicator allowing the application to signal overload (see `OverloadStatusIndicator`)
+* Added status indicator allowing the application to signal overload (see `LoadStatusIndicator`)
 
 ## Release 0.51.1
 * Updated to Spring Boot 1.3.3 (with Spring 4.2.5)